move: Select Screen under "generate" directory
