+++ app/src/main/java/com/ry05k2ulv/myapplication/ui/generate/select/SelectMaterialScreen.kt
@@ -1,4 +1,4 @@
-package com.ry05k2ulv.myapplication.ui.select
+package com.ry05k2ulv.myapplication.ui.generate.select
 
 import android.net.Uri
 import android.util.Log
