+++ app/src/main/java/com/ry05k2ulv/myapplication/ui/generate/select/SelectNavigation.kt
@@ -1,5 +1,6 @@
-package com.ry05k2ulv.myapplication.ui.select
+package com.ry05k2ulv.myapplication.ui.generate.select
 
+import android.net.Uri
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
@@ -11,10 +12,16 @@ fun NavController.navigateToSelect(navOptions: NavOptions?) {
     navigate(SELECT_NAVIGATION_ROUTE, navOptions)
 }
 
-fun NavGraphBuilder.selectScreen() {
+fun NavGraphBuilder.selectScreen(
+    onBack: () -> Unit,
+    onNext: () -> Unit,
+) {
     composable(
         route = SELECT_NAVIGATION_ROUTE
     ) {
-        SelectScreen()
+        SelectScreen(
+            onBack = onBack,
+            onNext = onNext
+        )
     }
 }
